refactor: 전체 코드 리팩토링4 _ mvc 패턴 수정
