+++ src/main/java/controller/InputController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import domain.Convert;
 import domain.Player;
 import view.InputView;
 
@@ -12,7 +13,7 @@ public class InputController {
 
     static List<String> getUserNames() {
         String names = InputView.inputPlayerNames(); // view 호출, 게임의 참여할 사람 이름 입력 받기
-        String[] userNames = Player.splitPlayerNames(names); // model 호출, 쉼표 기준으로 잘라서 배열에 저장
+        String[] userNames = Convert.splitNames(names); // model 호출, 쉼표 기준으로 잘라서 배열에 저장
         System.out.println("");
 
         return new ArrayList<>(Arrays.asList(userNames));
