+++ src/main/java/controller/PrintController.java
@@ -1,8 +1,8 @@
 package controller;
 
+import domain.Convert;
 import domain.Player;
-import domain.ResultStatistics;
-import domain.Validator;
+import domain.Result;
 import view.OutputView;
 
 import java.util.List;
@@ -33,12 +33,12 @@ public class PrintController {
 
     public static void printGameResult(List<Player> users, Player dealer) {
         printTotalPlayerResult(dealer, users); // 플레이어들의 최종 카드와 결과 출력
-        Map<String, String> userWinOrLoseResult = Validator.getUserWinOrLoseResult(dealer, users); // 유저 승패 결과 구하기
+        Map<String, String> userWinOrLoseResult = Result.getUserWinOrLoseResult(dealer, users); // 유저 승패 결과 구하기
         printWinOrLoseResult(userWinOrLoseResult); // 최종 승패 결과 출력 (딜러+유저)
     }
 
     private static void printPlayerInitialMessage(List<Player> users) {
-        String names = Player.getAddedCommaUserNames(users);
+        String names = Convert.getNamesWithComma(users);
 
         OutputView.printPlayerInitialMessage(names);
         System.out.println("");
@@ -47,8 +47,8 @@ public class PrintController {
     private static void printPlayerResult(Player player) {
         List<String> playersCards = player.cards;
 
-        int playersResult = ResultStatistics.getCardTotalSum(playersCards);
-        OutputView.printPlayerCardTotalResult(player, playersResult);
+        int playersResult = Result.getResultNumber(playersCards);
+        OutputView.printTotalPlayerResult(player, playersResult);
     }
 
     private static void printTotalPlayerResult(Player dealer, List<Player> users) {
@@ -62,7 +62,7 @@ public class PrintController {
 
     // 승패 결과 출력
     private static void printWinOrLoseResult(Map<String, String> userWinOrLoseResult) {
-        List<Integer> dealerResult = ResultStatistics.getDealerWinOrLoseResult(userWinOrLoseResult);
+        List<Integer> dealerResult = Result.getDealerWinOrLoseResult(userWinOrLoseResult);
 
         OutputView.printDealerWinOrLoseResult(dealerResult); // 딜러 승패 결과
         OutputView.printUsersWinOrLoseResult(userWinOrLoseResult); // 유저 승패 결과
